Add Create Group to PermissionManager. Also fix a bug in CaseContainer.
